requirements.txt
absl-py==1.0.0
aiobotocore==2.0.1
aiohttp==3.8.1
aioitertools==0.8.0
aiosignal==1.2.0
alembic==1.4.1
anyio==3.3.4
appnope==0.1.2
argon2-cffi==21.1.0
asgiref==3.4.1
astroid==2.8.6
astunparse==1.6.3
async-timeout==4.0.1
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
black==21.11b1
bleach==4.1.0
botocore==1.22.8
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cloudpickle==2.0.0
-e git+ssh://git@github.com/toporas/btc_sent_fc.git@2ae51321a67e4e46d8116871e0d79df1e2b32307#egg=CloudSentiment
colorama==0.4.4
coverage==6.2
cycler==0.11.0
Cython==0.29.24
databricks-cli==0.16.2
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
docker==5.0.3
docutils==0.18.1
entrypoints==0.3
fastapi==0.70.0
filelock==3.4.0
flake8==4.0.1
Flask==2.0.2
flatbuffers==2.0
fonttools==4.28.2
frozenlist==1.2.0
fsspec==2021.11.1
gast==0.4.0
gcsfs==2021.11.1
gitdb==4.0.9
GitPython==3.1.24
google-api-core==2.2.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-cloud-core==2.2.1
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
grpcio==1.42.0
gunicorn==20.1.0
h11==0.12.0
h5py==3.6.0
huggingface-hub==0.1.2
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
iniconfig==1.1.1
ipdb==0.13.9
ipykernel==6.5.1
ipython==7.29.0
ipython-genutils==0.2.0
isort==5.10.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
json5==0.9.6
jsonschema==4.2.1
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyter-server==1.11.2
jupyterlab==3.2.4
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
keras==2.7.0
Keras-Preprocessing==1.1.2
keyring==23.4.0
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
libclang==12.0.0
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mccabe==0.6.1
memoized-property==1.0.3
mistune==0.8.4
mlflow==1.22.0
multidict==5.2.0
mypy-extensions==0.4.3
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.6
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.9.0
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pkginfo==1.8.1
platformdirs==2.4.0
pluggy==1.0.0
pmdarima==1.8.4
prometheus-client==0.12.0
prometheus-flask-exporter==0.18.6
prompt-toolkit==3.0.22
protobuf==3.19.1
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.8.2
pyflakes==2.4.0
Pygments==2.10.0
pylint==2.11.1
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
python-editor==1.0.4
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
querystring-parser==1.2.4
readme-renderer==30.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.8
s3fs==2021.11.1
sacremoses==0.0.46
scikit-learn==1.0.1
scipy==1.7.2
seaborn==0.11.2
Send2Trash==1.8.0
setuptools-scm==6.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
SQLAlchemy==1.4.27
sqlparse==0.4.2
starlette==0.16.0
statsmodels==0.13.1
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tokenizers==0.10.3
toml==0.10.2
tomli==1.2.2
torch==1.10.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.12.5
twine==3.6.0
typing_extensions==4.0.0
urllib3==1.26.7
uvicorn==0.15.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.2
wrapt==1.13.3
yapf==0.31.0
yarl==1.7.2
zipp==3.6.0
